# knowsphere-ai-complete_timestamp/knowsphere-ai/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: KnowSphere AI API
  version: "6.0.0"
  description: |
    Enterprise Knowledge Intelligence Platform — REST API.

    Scoping note: this specification covers the primary request/response
    shape of every resource group (auth, providers, documents, chat,
    analytics, audit, observability, notifications, health). It documents
    the core, most-used parameters for each endpoint rather than every
    optional query parameter combination across 60+ endpoints — treat this
    as the authoritative contract for integration purposes, and the actual
    route source (`app/*/routes.py`) as the final word on any parameter
    not shown here.

    All endpoints are prefixed with `/api/v1`. Authentication is a JWT
    Bearer token in the `Authorization` header, obtained via `/auth/login`.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiSuccess:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: {}
        message: { type: string }
        request_id: { type: string, nullable: true }
    ApiError:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string, example: "VALIDATION_ERROR" }
            message: { type: string }
            details: { nullable: true }
        request_id: { type: string, nullable: true }
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    Forbidden:
      description: Authenticated, but role doesn't have permission
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
paths:
...
  /auth/login:
    post:
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        "200":
          description: Login successful — returns access and refresh tokens
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiSuccess" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { description: "Too many login attempts (brute-force rate limit, 10/min)" }
  /auth/refresh:
    post:
      summary: Rotate access/refresh tokens
      security: [{ bearerAuth: [] }]
      description: Requires the refresh token (not the access token) as the Bearer token.
      responses:
        "200": { description: New token pair issued }
...
  /auth/users:
    get:
      summary: List all users (admin only)
      responses:
        "200": { description: OK }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      summary: Provision a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, display_name]
              properties:
                email: { type: string }
                password: { type: string }
                display_name: { type: string }
                role: { type: string, enum: [admin, manager, employee] }
      responses:
        "201": { description: User created }
  /auth/users/{userId}:
    patch:
      summary: Update a user's role or active status (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string }
                is_active: { type: boolean }
      responses:
        "200": { description: Updated }
...